Added TomcatCustomizer for conditionally customizing tomcat based on existence of tomcat on the class path
