+++ src/main/resources/META-INF/spring.factories
@@ -1,5 +1,6 @@
 org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
   no.skatteetaten.aurora.config.ApplicationConfig,\
+  no.skatteetaten.aurora.TomcatCustomizer,\
   no.skatteetaten.aurora.config.AuroraMetricsConfig,\
   no.skatteetaten.aurora.AuroraMetrics
 org.springframework.context.ApplicationListener=no.skatteetaten.aurora.DatabasePropertiesListener
